chore: types
~ updated heartRate() function to accept the RGB values themselves,
  rather than enum for fixed values
~ changed temperature back to int type

+++ main/include/Motion.h
@@ -27,7 +27,7 @@ typedef struct {
         int16_t X, Y, Z;
     } Gyro;
 
-    double Temperature;
+    int16_t Temperature;
 } MotionValues;
